+++ aiobungie/crate/entity.py
@@ -28,28 +28,71 @@ This will include all Bungie Definitions.
 
 from __future__ import annotations
 
-__all__ = ("InventoryEntity", "Entity")
+__all__: tuple[str, ...] = (
+    "InventoryEntity",
+    "Entity",
+    "ObjectiveEntity",
+    "GatingScope",
+    "ValueUIStyle",
+    "BaseEntity",
+)
 
 import abc
 import typing
 
 import attr
 
+from aiobungie.internal import enums
+
 if typing.TYPE_CHECKING:
     from aiobungie import traits
+    from aiobungie import typedefs
     from aiobungie import undefined
     from aiobungie.internal import assets
-    from aiobungie.internal import enums
+
+
+@typing.final
+class GatingScope(enums.IntEnum):
+    """An enum represents restrictive type of gating that is being performed by an entity.
+
+    This is useful as a shortcut to avoid a lot of lookups when determining whether the gating on an Entity
+    applies to everyone equally, or to their specific Profile or Character states.
+    """
+
+    NONE = 0
+    GLOBAL = 1
+    CLAN = 2
+    PROFILE = 3
+    CHARACTER = 4
+    ITEM = 5
+    ASSUMED_WORST_CASE = 6
+
+
+@typing.final
+class ValueUIStyle(enums.IntEnum):
+    AUTOMATIC = 0
+    FRACTION = 1
+    CHECK_BOX = 2
+    PERCENTAGE = 3
+    DATETIME = 4
+    FRACTION_FLOAT = 5
+    INTEGER = 6
+    TIME_DURATION = 7
+    HIDDEN = 8
+    MULTIPLIER = 9
+    RED_PIPS = 11
+    EXPLICIT_PERCENTAGE = 12
+    RAW_FLOAT = 13
 
 
 class Entity(abc.ABC):
-    """An interface of a Bungie Definition Entity.
+    """An interface of any Bungie Definition/Entity.
 
     This is the main entity which all other entities should inherit from.
     it holds core information that all bungie entities has.
     """
 
-    __slots__: typing.Sequence[str] = ()
+    __slots__ = ()
 
     @property
     @abc.abstractmethod
@@ -93,33 +136,42 @@ class Entity(abc.ABC):
         return self.hash
 
 
+@attr.mutable(kw_only=True, weakref_slot=False)
+class BaseEntity(Entity):
+    """Concerate Bungie entity implementation."""
+
+    # These are not attribs on purpose.
+    # We dont want to redefine them again in the actual entity
+    # implementation.
+
+    net: traits.Netrunner = attr.field(repr=False)
+    # <<inherited docstring from Entity>>.
+
+    hash: int
+    # <<inherited docstring from Entity>>.
+
+    index: int
+    # <<inherited docstring from Entity>>.
+
+    name: undefined.UndefinedOr[str]
+    # <<inherited docstring from Entity>>.
+
+    description: undefined.UndefinedOr[str]
+    # <<inherited docstring from Entity>>.
+
+    icon: assets.MaybeImage
+    # <<inherited docstring from Entity>>.
+
+    has_icon: bool
+    # <<inherited docstring from Entity>>.
+
+
 @attr.define(kw_only=True, hash=False, weakref_slot=False)
-class InventoryEntity(Entity):
+class InventoryEntity(BaseEntity, Entity):
     """Represents a bungie inventory item entity.
 
     This derives from `DestinyInventoryItemDefinition` definition.
     """
-
-    net: traits.Netrunner = attr.field(repr=False, hash=False, eq=False)
-    """A network state used for making external requests."""
-
-    hash: int = attr.field(repr=True, hash=True, eq=True)
-    """Entity's hash."""
-
-    index: int = attr.field(repr=True, hash=False, eq=False)
-    """Entity's index."""
-
-    name: undefined.UndefinedOr[str] = attr.field(repr=True, hash=False, eq=False)
-    """Entity's name. This can be `UNDEFINED` if not found."""
-
-    description: undefined.UndefinedOr[str] = attr.field(repr=True)
-    """Entity's description."""
-
-    icon: assets.MaybeImage = attr.field(repr=False, hash=False, eq=False)
-    """Entity's icon"""
-
-    has_icon: bool = attr.field(repr=False, hash=False, eq=False)
-    """A boolean that returns True if the entity has an icon."""
 
     type: undefined.UndefinedOr[enums.Item] = attr.field(repr=True, hash=False)
     """Entity's type. Can be undefined if nothing was found."""
@@ -141,7 +193,7 @@ class InventoryEntity(Entity):
     bucket_type: typing.Optional[int] = attr.field(repr=True, hash=False, eq=False)
     """The entity's bucket type, None if unknown"""
 
-    stats: typing.Optional[typing.Dict[str, typing.Any]] = attr.field(
+    stats: typing.Optional[dict[str, typing.Any]] = attr.field(
         repr=False, hash=False, eq=False
     )
     """Entity's stats. this currently returns a dict object
@@ -186,3 +238,47 @@ class InventoryEntity(Entity):
 
     banner: typing.Optional[assets.Image] = attr.field(repr=False, eq=False, hash=False)
     """Entity's banner."""
+
+
+@attr.define(kw_only=True, weakref_slot=False)
+class ObjectiveEntity(BaseEntity, Entity):
+    """Represents a bungie inventory item entity.
+
+    This derives from `DestinyObjectiveDefinition` definition.
+    """
+
+    # TODO: document these.
+
+    unlock_value_hash: int = attr.field()
+
+    minimum_visibility: int = attr.field()
+
+    completion_value: int = attr.field()
+
+    scope: typedefs.IntAnd[GatingScope] = attr.field()
+
+    location_hash: int = attr.field()
+
+    allowed_negative_value: bool = attr.field()
+
+    allowed_value_change: bool = attr.field()
+
+    counting_downward: bool = attr.field()
+
+    display_only_objective: bool = attr.field()
+
+    value_style: typedefs.IntAnd[ValueUIStyle] = attr.field()
+
+    complete_value_style: typedefs.IntAnd[ValueUIStyle] = attr.field()
+
+    progress_value_style: typedefs.IntAnd[ValueUIStyle] = attr.field()
+
+    allow_over_completion: bool = attr.field()
+
+    show_value_style: typedefs.IntAnd[ValueUIStyle] = attr.field()
+
+    progress_description: str = attr.field()
+
+    perks: dict[str, int] = attr.field()
+
+    stats: dict[str, int] = attr.field()